Complete board relayout needed if we want the Daughter card to fit in the student version of EAGLE, see desc.
Currently, the board size that is allowed by EAGLE student is just big
enough to accomodate the daughter card. However, the current orientation
of the daughter card means that, in order to place the shutdown parts on
the correct part of the daughter card, the sbRIO will have to expand in
a direction that it doesn't like. As such, we need to re lay out the
sbrio so that it can